Move consts to parent class
